@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import "./App.scss";
 import About from "./Components/About/About";
 import Contact from "./Components/Contact/Contact";
