@@ -1,14 +1,14 @@
-import React from "react";
 import { Tooltip } from "react-tooltip";
 import SectionHeader from "../SectionHeader/SectionHeader";
-import SkillIcon from "../SkillIcon/SkillIcon";
-import { gitHubSvg, emailSvg, instaSvg } from "../../data/skillsData";
+import Icon from "../Icon/Icon";
+import Email from '../../Components/SvgIconPaths/Email';
+import Github from '../../Components/SvgIconPaths/Github';
+import Instagram from '../../Components/SvgIconPaths/Instagram';
 import styles from "./Contact.module.scss";
 
 const Contact = () => {
   const urls = {
     github: "http://github.com/tonylubin",
-    linkedin: "http://linkedin.com",
     email: "anthonylubin@outlook.com",
     instagram: "https://www.instagram.com/anthonylubin78"
   };
@@ -19,7 +19,7 @@ const Contact = () => {
       <div className={styles.contactSection__container}>
         <div className={styles.emailHeading}>
           <h3>Email:</h3>
-          <SkillIcon width="2em" height="2em" viewBox={emailSvg.viewBox} svgInfo={emailSvg.svgInfo} /> 
+          <Icon classname={styles.iconSvg} path={<Email />} /> 
           <a
             className={styles.emailHeading__link}
             href={`mailto:${urls.email}`}
@@ -36,25 +36,11 @@ const Contact = () => {
             data-tooltip-id="gh"
             data-tooltip-content="Vist my Github page"
             data-tooltip-place="bottom"
-            data-tooltip-variant="info"
+            data-tooltip-variant="dark"
           >
-            <SkillIcon width="2em" height="2em" viewBox={gitHubSvg.viewBox} svgInfo={gitHubSvg.svgInfo}/>
+            <Icon classname={styles.iconSvg} path={<Github />} />
           </a>
           <Tooltip id="gh" />
-          {/* <a
-            className={styles.linkHover}
-            href={urls.linkedin}
-            target="_blank"
-            rel="noreferrer"
-            data-tooltip-id="lkdn"
-            data-tooltip-content="Vist my LinkedIn page"
-            data-tooltip-place="bottom"
-            data-tooltip-variant="info"
-          >
-            <SkillIcon width="2em" height="2em" viewBox={linkedInSvg.viewBox} svgInfo={linkedInSvg.svgInfo}/>
-          </a>
-          <Tooltip id="lkdn" />
-         */}
           <a
             className={styles.linkHover}
             href={urls.instagram}
@@ -63,9 +49,9 @@ const Contact = () => {
             data-tooltip-id="insta"
             data-tooltip-content="Take a peek at my Insta"
             data-tooltip-place="bottom"
-            data-tooltip-variant="info"
+            data-tooltip-variant="dark"
           >
-            <SkillIcon width="2em" height="2em" viewBox={instaSvg.viewBox} svgInfo={instaSvg.svgInfo}/>
+            <Instagram classname={styles.iconSvg} />
           </a>
           <Tooltip id="insta" />
         </div>
