@@ -1,4 +1,3 @@
-import React from "react";
 import TypewriterComponent from "typewriter-effect";
 import styles from "./LandingPage.module.scss";
 
@@ -9,10 +8,10 @@ const LandingPage = () => {
         <TypewriterComponent
           onInit={(typewriter) => {
             typewriter.pauseFor(50).changeDelay(85)
-            .typeString("<span style='font-weight: 700; font-size: clamp(1.9rem, 5vw, 2.5rem); color: #c3dcaf;'>Hello</span>, (tech world),</br>")
-            .typeString("<span style='font-weight: 700'; color: #c3dcaf;'>I'm Anthony Lubin, </span>a junior web developer based in London. </br>")
-            .typeString("</br><span style='font-weight: 700; color: #c3dcaf;'>Please </span>scroll down to</br>")
-            .typeString("<span style='font-weight: 700; color: #c3dcaf;'>Enter</span>.....")
+            .typeString("<span style='font-weight: 700; font-size: clamp(1.9rem, 5vw, 2.5rem);'>Hello</span>, (tech world),</br>")
+            .typeString("<span style='font-weight: 700';'>I'm Anthony Lubin, </span>a web developer based in London. </br>")
+            .typeString("</br><span style='font-weight: 700;'>Please </span>scroll down to</br>")
+            .typeString("<span style='font-weight: 700;'>Enter</span>.....")
             .start();
           }}        
         />
