@@ -1,11 +1,9 @@
-import React from 'react';
 import styles from './SkillsCard.module.scss';
 
-
-const SkillsCard = ({ icon, title }) => {
+const SkillsCard = ({ icon, title, tooltip }) => {
 
   return (
-  <div className={styles.card}>
+  <div className={styles.card} title={tooltip}>
       {icon}
       <h5 className={styles.card__title}>{title}</h5>
   </div>
