@@ -1,24 +1,31 @@
-import React from "react";
 import styles from "./SkillsTechList.module.scss";
-import SkillIcon from "../SkillIcon/SkillIcon";
+import Icon from "../Icon/Icon";
 import SkillsCard from "../SkillsCard/SkillsCard";
+import { techSkills } from "../../data/skillsData";
 
 const SkillsTechList = ({ techstackIcons }) => {
 
-  const getSkillsTechStack = techstackIcons.map((techstack, index) => (
+  const getSkillsTechStack = techstackIcons.map((skill, i) => {
 
-        <SkillsCard 
-        key = {index}
-        icon = {<SkillIcon svgInfo={techstack.svgInfo} viewBox={techstack.viewBox} width="2em" height="2em"/>}
-        />
+    const iconIndex = techSkills.findIndex((tech) => tech.title.toLowerCase() === skill);
+  
+    const techIcon = techSkills[iconIndex];
+    
+    return (
+      <SkillsCard
+        key={i}
+        tooltip={techIcon.title}
+        icon={
+          <Icon
+            path={techIcon.icon}
+            classname={styles.skillIcon}
+          />
+        }
+      />
+    )
+  });
 
-  ));    
-
-  return (
-    <div className={styles.skillsList}>
-      {getSkillsTechStack}
-    </div>
-  );
+  return <div className={styles.skillsList}>{getSkillsTechStack}</div>;
 };
 
 export default SkillsTechList;
